WIN32 VS fixes:  precompiler checking if size_t == unsigned int to avoid re-declaration  numeric constant instead of limits::max  fix -(unsigned int)  define M_PI
add missing references: SDL_ttf, ANN in KDE

+++ sdl_example/SpikeDetectorProcessor.cpp
@@ -52,7 +52,7 @@ SpikeDetectorProcessor::SpikeDetectorProcessor(LFPBuffer* buffer, const char* fi
     filt_pos = buffer->HEADER_LEN;
     
     for (size_t t=0; t < buffer->tetr_info_->tetrodes_number(); ++t) {
-        buffer->last_spike_pos_[t] = - refractory_;
+        buffer->last_spike_pos_[t] = - (int)refractory_;
     }
     
     //printf("filt len: %d\n", filter_len);
